# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music API",
    "description": "A comprehensive REST API for managing artists, albums, songs, and playlists with full CRUD operations, authentication, and data validation",
    "version": "2.0.0",
    "contact": {
      "name": "Music API Support",
      "email": "support@musicapi.com"
    }
  },
  "servers": [
    {
      "url": "https://music-api-cse341-u669.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/playlists": {
      "get": {
        "tags": ["Playlists"],
        "summary": "Get all playlists with filtering and pagination",
        "parameters": [
          {
            "name": "creator",
            "in": "query",
            "description": "Filter by creator name",
            "schema": { "type": "string" }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Filter by tag",
            "schema": { "type": "string" }
          },
          {
            "name": "is_public",
            "in": "query",
            "description": "Filter by public status",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": { "description": "Successfully retrieved playlists" }
        }
      },
      "post": {
        "tags": ["Playlists"],
        "summary": "Create a new playlist",
        "security": [{ "OAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlaylistInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Playlist created successfully" },
          "401": { "description": "Authentication required" }
        }
      }
    },
    "/playlists/{id}": {
      "get": {
        "tags": ["Playlists"],
        "summary": "Get playlist by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Playlist found" },
          "404": { "description": "Playlist not found" }
        }
      },
      "put": {
        "tags": ["Playlists"],
        "summary": "Update playlist",
        "security": [{ "OAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlaylistInput" }
            }
          }
        },
        "responses": {
          "204": { "description": "Playlist updated successfully" },
          "401": { "description": "Authentication required" }
        }
      },
      "delete": {
        "tags": ["Playlists"],
        "summary": "Delete playlist",
        "security": [{ "OAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Playlist deleted successfully" },
          "401": { "description": "Authentication required" }
        }
      }
    },
    "/playlists/{id}/songs": {
      "get": {
        "tags": ["Playlists"],
        "summary": "Get all songs in a playlist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Successfully retrieved playlist songs" },
          "404": { "description": "Playlist not found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PlaylistInput": {
        "type": "object",
        "required": ["name", "creator_name"],
        "properties": {
          "name": { "type": "string", "example": "My Favorite Songs" },
          "creator_name": { "type": "string", "example": "John Doe" },
          "description": { "type": "string", "example": "A collection of my favorite tracks" },
          "songs": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["60f7c2b8d4f1c12345678901", "60f7c2b8d4f1c12345678902"]
          },
          "tags": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["rock", "favorites", "classic"]
          },
          "is_public": { "type": "boolean", "example": true },
          "cover_image_url": { "type": "string", "example": "https://example.com/playlist.jpg" }
        }
      }
    },
    "securitySchemes": {
      "OAuth": {
        "type": "oauth2",
        "description": "OAuth 2.0 authentication with Google",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/auth/google",
            "tokenUrl": "/auth/google/callback",
            "scopes": {
              "profile": "Access user profile",
              "email": "Access user email"
            }
          }
        }
      }
    }
  }
}